Moved class "Drawing" into drawing.js
